+++ sample/src/jvmMain/kotlin/util/WatchableMapSerializer.kt
@@ -19,22 +19,24 @@ package util
 import io.gladed.watchable.WatchableList
 import io.gladed.watchable.WatchableMap
 import io.gladed.watchable.toWatchableMap
-import kotlinx.serialization.Decoder
-import kotlinx.serialization.Encoder
+import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.KSerializer
-import kotlinx.serialization.SerialDescriptor
 import kotlinx.serialization.Serializer
 import kotlinx.serialization.builtins.MapSerializer
 import kotlinx.serialization.builtins.serializer
-import kotlinx.serialization.mapDescriptor
+import kotlinx.serialization.descriptors.buildClassSerialDescriptor
+import kotlinx.serialization.descriptors.mapSerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
 
+@OptIn(ExperimentalSerializationApi::class)
 @Serializer(forClass = WatchableList::class)
 class WatchableMapSerializer<T : Any>(
     stringSerializer: KSerializer<String>,
     valueSerializer: KSerializer<T>
 ) : KSerializer<WatchableMap<String, T>> {
-    override val descriptor = SerialDescriptor("WatchableMap") {
-        mapDescriptor(stringSerializer.descriptor, valueSerializer.descriptor)
+    override val descriptor = buildClassSerialDescriptor("WatchableMap") {
+        mapSerialDescriptor(stringSerializer.descriptor, valueSerializer.descriptor)
     }
 
     private val mapSerializer = MapSerializer(String.serializer(), valueSerializer)