+++ sample/src/jvmMain/kotlin/util/util.kt
@@ -21,14 +21,12 @@ import kotlinx.serialization.json.Json
 import io.gladed.watchable.store.Transformer
 import io.gladed.watchable.store.Store
 import io.gladed.watchable.store.transform
-import kotlinx.serialization.UnstableDefault
 
 /** Convert this [KSerializer] to an [Transformer] of [String] and [T] */
-@OptIn(UnstableDefault::class)
 fun <T : Any> KSerializer<T>.toInflator() = object : Transformer<String, T> {
     @Suppress("DEPRECATION")
-    override fun toTarget(value: String) = Json.nonstrict.parse(this@toInflator, value)
-    override fun fromTarget(value: T): String = Json.stringify(this@toInflator, value)
+    override fun toTarget(value: String) = Json.decodeFromString(this@toInflator, value)
+    override fun fromTarget(value: T): String = Json.encodeToString(this@toInflator, value)
 }
 
 fun <T : Any> Store<String>.serialize(serializer: KSerializer<T>): Store<T> = transform(serializer.toInflator())